currUserData at globle
